refactored zping to use Cobra cli
