+++ example/zping/go.mod
@@ -1,0 +1,13 @@
+module zping
+
+go 1.16
+
+require (
+	github.com/michaelquigley/pfxlog v0.6.1
+	github.com/openziti/foundation v0.15.76 // indirect
+	github.com/openziti/sdk-golang v0.15.96
+	github.com/pkg/errors v0.9.1
+	github.com/sirupsen/logrus v1.8.1 // indirect
+	github.com/spf13/cobra v1.2.1
+	github.com/spf13/viper v1.8.1
+)
